requirements.txt
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastcore==1.3.19
idna==2.10
iniconfig==1.1.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
matplotlib==3.4.1
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pathspec==0.8.1
pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
seaborn==0.11.1
six==1.15.0
streamlit==0.81.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
